Правка архитектуры
