+++ prisma/schema.prisma
@@ -14,34 +14,6 @@ model User {
   password  String
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
-  posts     Post[]
-}
-
-model Post {
-  id        Int         @id @default(autoincrement())
-  title     String
-  content   String
-  author    User        @relation(fields: [authorId], references: [id])
-  authorId  Int
-  createdAt DateTime    @default(now())
-  updatedAt DateTime    @updatedAt
-  posts     PostOnTag[]
-}
-
-model Tag {
-  id    Int         @id @default(autoincrement())
-  name  String
-  posts PostOnTag[]
-}
-
-model PostOnTag {
-  id        Int      @id @default(autoincrement())
-  post      Post     @relation(fields: [postId], references: [id])
-  postId    Int
-  tag       Tag      @relation(fields: [tagId], references: [id])
-  tagId     Int
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
 }
 
 model Products {
@@ -53,8 +25,8 @@ model Products {
   updatedAt   DateTime @updatedAt
 }
 
-// model Location {
-//   id      Int    @id @default(autoincrement())
-//   city    String
-//   address String
-// }
+model Location {
+  id      Int    @id @default(autoincrement())
+  city    String
+  address String
+}
