+++ src/app.ts
@@ -1,8 +1,11 @@
 import { PrismaClient } from '@prisma/client';
-import { Markup, Scenes, Telegraf } from 'telegraf';
+import { Markup, Scenes } from 'telegraf';
 import 'dotenv/config';
 import { MyContext } from './Types/context';
 import LocalSession from 'telegraf-session-local';
+
+import { bot } from './Bot/bot.service';
+import { addressScene, cityScene } from './Scenes/locationScenes';
 const prisma = new PrismaClient();
 
 export class App {
@@ -10,43 +13,6 @@ export class App {
 		await prisma.$connect();
 	}
 }
-
-const token = process.env.TOKEN;
-if (!token) {
-	throw new Error('Не задан token');
-}
-
-export const cityScene = new Scenes.BaseScene<MyContext>('city');
-
-cityScene.enter((ctx) => ctx.reply('Укажите свой город'));
-
-cityScene.on('text', (ctx) => {
-	const currCity = String(ctx.message.text);
-	if (currCity) {
-		ctx.reply('Спасибо за Город!!!');
-		ctx.scene.enter('address');
-	} else {
-		ctx.reply('Укажите город строкой!!!');
-		ctx.scene.reenter();
-	}
-});
-cityScene.on('message', (ctx) => ctx.reply('Давай лучше город!'));
-
-export const addressScene = new Scenes.BaseScene<MyContext>('address');
-addressScene.enter((ctx) => ctx.reply('Укажите свой адрес'));
-addressScene.on('text', (ctx) => {
-	const currAddress = String(ctx.message.text);
-	if (currAddress) {
-		ctx.reply('Спасибо за Адрес!!!');
-		ctx.scene.leave();
-	} else {
-		ctx.reply('Укажите город строкой!!!');
-		ctx.scene.reenter();
-	}
-});
-addressScene.on('message', (ctx) => ctx.reply('Давай лучше адрес!'));
-
-const bot = new Telegraf<MyContext>(token);
 
 const stage = new Scenes.Stage<MyContext>([cityScene, addressScene]);
 const { leave, enter } = Scenes.Stage;
