+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "ES6",
     "experimentalDecorators": true,
-    "module": "ES6",
+    "module": "CommonJS",
     "moduleResolution": "node",
     "outDir": "./dist",
     "removeComments": true,
